home
home

+++ categories.js
@@ -8,11 +8,12 @@ const jwt = require("jsonwebtoken");
 dotenv.config();
 
 
-const router = express.Router();
 
+const router = express.Router();
+router.use(express.json())
 
 const verifyToken = ((req, res, next) => {
-    const authHeader = req.headers["authorization"];
+    const authHeader = req.headers['authorization'];
     const token = authHeader && authHeader.split(' ')[1];
     if (token == null) return res.status(401).json("Token null");
 
@@ -23,7 +24,7 @@ const verifyToken = ((req, res, next) => {
     })
 })
 
-router.get("/categories", verifyToken, async (req, res) => {
+router.get("/categories", async (req, res) => {
     const categoriesData = await sql`SELECT * FROM categories;`
     res.send(categoriesData);
 })
@@ -31,11 +32,12 @@ router.get("/categories", verifyToken, async (req, res) => {
 router.post("/categories", verifyToken, async (req, res) => {
     const { user_id, name, category_image } = req.body;
     try {
+        console.log('user', user_id)
         console.log('Name:', name);
         console.log('Category Image:', category_image);
 
-        const categories = await sql`INSERT INTO categories (user_id,name, createdAt, updatedAt, category_image)
-            VALUES (${user_id},${name}, NOW(), NOW(), ${category_image})`;
+        const categories = await sql`INSERT INTO categories (user_id, name, createdAt, updatedAt, category_image)
+            VALUES (${user_id}, ${name}, NOW(), NOW(), ${category_image})`;
 
         res.status(201).json({ success: true, data: categories[0] });
     } catch (error) {
@@ -45,24 +47,25 @@ router.post("/categories", verifyToken, async (req, res) => {
 
 })
 
-// router.delete("/categories/:categoryId", async (req, res) => {
-//     const categoryId = req.params.categoryId;
+router.delete("/categories/:categoryId", verifyToken, async (req, res) => {
+    const categoryId = req.params.categoryId;
+    console.log('id', categoryId);
+    const existingCategory = await sql`SELECT * FROM categories WHERE id = ${categoryId}`;
+    if (!existingCategory || existingCategory.length === 0) {
+        return res.status(404).json({ success: false, error: "Category not found" })
+    }
 
-//     const existingCategory = await sql`SELECT * FROM categories WHERE id = ${categoryId}`;
-//     if (!existingCategory || existingCategory.length === 0) {
-//         return res.status(404).json({ success: false, error: "Category not found" })
-//     }
-
-//     await sql`DELETE FROM categories WHERE id=${categoryId}`
-//     return res.status(202).json({ success: true, message: "Successfully deleted" })
-// })
+    await sql`DELETE FROM categories WHERE id=${categoryId}`
+    return res.status(202).json({ success: true, message: "Successfully deleted" })
+})
 
 router.delete("/categories", verifyToken, async (req, res) => {
+    const user_id = req.user.userId;
     try {
-        await sql`DELETE FROM categories`
+        await sql`DELETE FROM categories WHERE user_id=${user_id}`
         res.status(202).json({ success: true, message: "Deleted all categories" })
     } catch (error) {
-        res.status(505).json({ success: false, error: "Cannot delete" })
+        res.status(400).json({ success: false, error: "Cannot delete" })
     }
 })
